src/backend/pyproject.toml
[tool.poetry]
name = "catalog"
version = "0.1.0"
description = ""
authors = ["ykozhanov <ykozhanov97@gmail.com>"]

[tool.poetry.dependencies]
python = "3.13.0"
sqlalchemy = "2.0.36"
pydantic = "2.10.4"
flask = "3.1.0"
psycopg2 = "2.9.10"
python-dotenv = "1.0.1"
bcrypt = "4.2.1"
pyjwt = "2.10.1"
email-validator = "2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ["sqlalchemy.ext.mypy.plugin"]

